# api/openapi.json
{
    "openapi":"3.0.2",
    "info": {
      "title":"Province of BC Web Applications API",
      "version":"1.0",
      "description": "This API enables tracking and management of web applications developed by the Province of BC."
    },
    "servers": [
      {"url":"http://localhost:3000"}
    ],
    "components": {
        "schemas":{
            "product":{
                "type": "object",
                "properties":{
                    "productId":{
                        "type": "integer",
                        "description": "ID of product"
                    },
                    "productName":{
                        "type": "string",
                        "description": "Name of product"
                    },
                    "productOwnerName":{
                        "type": "string",
                        "description": "Name of Product Manager who owns product"
                    },
                    "Developers":{
                        "type": "array",
                        "description": "List of developers who are working on product"
                    },
                    "scrumMasterName":{
                        "type": "string",
                        "description": "Name of Program Manager who is overseeing product"
                    },
                    "startDate":{
                        "type": "string",
                        "description": "Date product was started in 'YYYY/MM/DD' format"
                    },
                    "methodology":{
                        "type": "string",
                        "description": "Project management methodology being used for project"
                    },
                    "location":{
                        "type":"string",
                        "description": "URL for GitHub repository of product"
                    }
                },
                "example":{
                    "productId": 5,
                    "productName": "obtuse_blue_iguana",
                    "productOwnerName": "Alex",
                    "Developers": ["Erin","Lin","Yusuf","Ajay","Charles"],
                    "scrumMasterName": "Mircea",
                    "startDate": "2022/10/09",
                    "methodology": "Agile",
                    "location": "github.com/bcgov/obtuse_blue_iguana"
                }
            },
            "productArray":{
                "type": "array",
                "items":{
                    "$ref": "#/components/schemas/product"
                }
            }
        }
    },
    "paths": {
        "/api/product/all": {
            "get": {
                "description": "Fetches all products",
                "tags": ["Product"],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content":{
                            "application/json": {
                                "schema": {
                                        "$ref": "#/components/schemas/productArray"
                                    }
                                }
                            }
                        }   
                    }
                }
            },
        "/api/health": {
            "get": {
                "description": "Checks current health of API - currently only includes a check for existence of mock DB file",
                "tags": ["Monitoring"],
                "responses": {
                    "200": {
                        "description": "API is up and mock DB file is present"
                    },
                    "503": {
                        "description": "Mock db does not exist. Run generated-db.js script to create mock db json file"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/product": {
            "post":{
                "description": "Creates new product record",
                "tags": ["Product"],
                "requestBody":{
                    "required": true,
                    "content": {
                        "application/json":{
                            "schema":{
                                "$ref": "#/components/schemas/product"
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description": "OK"
                    },
                    "500":{
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/product/{productId}":{
            "get":{
                "description": "Retrieves single product corresponding to productId provided in path",
                "tags": ["Product"],
                "parameters":[{
                    "in": "path",
                    "name": "productId",
                    "schema": {
                        "type": "integer"
                    },
                    "required": true
                }],
                "responses": {
                    "200":{
                        "description": "OK",
                        "content":{
                            "application/json":{
                                "schema": {
                                    "$ref": "#/components/schemas/product"
                                }
                            }
                        }
                    },
                    "400":{
                        "description": "productId is not an integer"
                    },
                    "404":{
                        "description": "productId is not found within mock db"
                    },
                    "500":{
                        "description": "Server error"
                    }
                }
            },
            "delete":{
                "description": "Deletes single product corresponding to productId provided in path",
                "tags": ["Product"],
                "parameters":[{
                    "in": "path",
                    "name": "productId",
                    "schema": {
                        "type": "integer"
                    },
                    "required": true
                }],
                "responses":{
                    "204":{
                        "description": "Deletion completed - no content"
                    },
                    "404": {
                        "description": "productId is not found within mock db"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "put": {
                "description": "Overwrites existing product record with passed values",
                "tags": ["Product"],
                "requestBody":{
                    "content": {
                        "application/json":{
                            "schema":{
                                "$ref": "#/components/schemas/product"
                            }
                        }
                    }
                },
                "parameters": [{
                    "in": "path",
                    "name": "productId",
                    "schema":{
                        "type": "integer"
                    },
                    "required": true
                }],
                "responses": {
                    "200":{
                        "description": "Update completed"
                    },
                    "404":{
                        "description": "productId not found within mock db"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }  
    }
}